+++ changelog.md
@@ -4,7 +4,10 @@ title: Changelog
 
 
 ### v0.1.0
+
 - 2023-12-09
   - [x] Create `changelog.md`.
   - [x] Enable to show images in Markdown.
   - [x] Add table of contents to accordion menu.
+- 2023-12-13
+  - [x] Enable to show thumbnail in *Blog* menu.